Refactor generators into smaller modules
Use constants instead of strings in all  and refactored endpoints.
The constants should help with maintainability as they can be changed,
tracked and refactored more easily than to string literals.
