+++ muse_for_anything/api/v1_api/generators/__init__.py
@@ -1,0 +1,1 @@
+"""Module containing link, api key and resource generators."""
